Twitter2-main/server/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
bcrypt==4.1.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decorator==4.4.2
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
facenet-pytorch==2.6.0
fastapi==0.109.2
fer==22.5.1
ffmpeg==1.4
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
keras==3.9.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
moviepy==1.0.3
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pillow==10.2.0
proglog==0.1.10
protobuf==5.29.3
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.19.1
PyJWT==2.8.0
pymongo==4.6.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.9
pytz==2025.1
requests==2.32.3
rich==13.9.4
setuptools==76.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.36.3
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.27.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2